Minor fixes

+++ include/data_augmentation/random_crop.hpp
@@ -48,7 +48,7 @@ private:
 
   void apply_crop(Tensor<T> &data, size_t batch_idx, size_t channels, size_t height, size_t width,
                   int start_x, int start_y) {
-    const int padded_size = width + 2 * padding_;
+    const size_t padded_size = width + 2 * padding_;
     Tensor<T> padded({1, channels, padded_size, padded_size});
     padded.fill(static_cast<T>(0));
